added old dsp.in tests

+++ src/audio3d/dsp_out.py
@@ -21,11 +21,11 @@ class DspOut:
         # Number of all speakers
         self.spn = len(self.state.gui_sp)
         self.sp_binaural_block = [np.zeros((
-            fft_blocksize, 2), dtype=np.float64) for sp in range(self.spn)]
-        self.sp_binaural_block_out = [np.zeros((hopsize, 2), dtype=np.float64)
+            fft_blocksize, 2), dtype=np.float32) for sp in range(self.spn)]
+        self.sp_binaural_block_out = [np.zeros((hopsize, 2), dtype=np.float32)
                                       for sp in range(self.spn)]
         self.sp_binaural_block_add = [np.zeros((fft_blocksize - hopsize, 2),
-                                      dtype=np.float64) for sp in range(
+                                      dtype=np.float32) for sp in range(
             self.spn)]
         self.binaural_block = np.zeros((hopsize, 2), dtype=np.float32)
         self.continue_convolution = [True for sp in range(self.spn)]
